Refactored WorldGuard variables into its own class
