+++ src/main/java/plugins/nate/smp/SMP.java
@@ -1,9 +1,5 @@
 package plugins.nate.smp;
 
-import com.sk89q.worldguard.WorldGuard;
-import com.sk89q.worldguard.protection.flags.StateFlag;
-import com.sk89q.worldguard.protection.flags.registry.FlagConflictException;
-import com.sk89q.worldguard.protection.flags.registry.FlagRegistry;
 import net.coreprotect.CoreProtectAPI;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -19,9 +15,6 @@ import java.util.logging.Logger;
 public final class SMP extends JavaPlugin {
     private static SMP plugin;
     private static CoreProtectAPI coreProtect;
-
-    public static StateFlag WITHER_EXPLOSIONS;
-    public static StateFlag BANK_FLAG;
 
     public static final Logger logger = Logger.getLogger("Minecraft");
     public final File prefixesFile = new File(getDataFolder() + "/prefixes.yml");
@@ -48,18 +41,7 @@ public final class SMP extends JavaPlugin {
     @Override
     public void onLoad() {
         super.onLoad();
-
-        FlagRegistry registry = WorldGuard.getInstance().getFlagRegistry();
-        try {
-            StateFlag witherExplosionsFlag = new StateFlag("wither-explosions", true);
-            StateFlag bankFlag = new StateFlag("bank", false);
-
-            registry.register(witherExplosionsFlag);
-            registry.register(bankFlag);
-
-            WITHER_EXPLOSIONS = witherExplosionsFlag;
-            BANK_FLAG = bankFlag;
-        } catch (FlagConflictException ignored) {}
+        WorldGuardUtils.registerFlags();
     }
 
     public static SMP getPlugin() {
